# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TGS Admin API Documentation",
        "description": "Documentation de l'API REST pour le système TGS Admin - Toulouse Game Show",
        "contact": {
            "email": "support@tgs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://votre-domaine.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/v1/salons/{salon}/ticket-prices": {
            "get": {
                "tags": [
                    "Ticket Prices"
                ],
                "summary": "Liste des prix de billets",
                "description": "Récupère tous les prix de billets d'un salon avec leurs contenus",
                "operationId": "getSalonTicketPrices",
                "parameters": [
                    {
                        "name": "salon",
                        "in": "path",
                        "description": "ID du salon",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prix de billets récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Pass 1 jour"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": 15
                                                    },
                                                    "contents": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "ticket_contents": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/salons/{salon}/ticket-prices/{ticketPrice}": {
            "get": {
                "tags": [
                    "Ticket Prices"
                ],
                "summary": "Détail d'un prix de billet",
                "description": "Récupère les informations d'un prix de billet spécifique avec ses contenus",
                "operationId": "getSalonTicketPrice",
                "parameters": [
                    {
                        "name": "salon",
                        "in": "path",
                        "description": "ID du salon",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "ticketPrice",
                        "in": "path",
                        "description": "ID du prix de billet",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prix de billet récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "number"
                                                },
                                                "contents": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "ticket_contents": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Prix de billet non trouvé"
                    }
                }
            }
        }
    },
...
}